+++ packages/jest-diff/src/printDiffs.ts
@@ -7,32 +7,28 @@
 
 import chalk from 'chalk';
 
-import {
-  DIFF_DELETE,
-  DIFF_EQUAL,
-  DIFF_INSERT,
-  Diff,
-  cleanupSemantic,
-} from './cleanupSemantic';
+import {DIFF_EQUAL, Diff, cleanupSemantic} from './cleanupSemantic';
+import diffLines from './diffLines';
 import diffStrings from './diffStrings';
 import getAlignedDiffs from './getAlignedDiffs';
 import {
   joinAlignedDiffsExpand,
   joinAlignedDiffsNoExpand,
 } from './joinAlignedDiffs';
-import {DiffOptions} from './types';
+import {normalizeDiffOptions} from './normalizeDiffOptions';
+import {DiffOptions, DiffOptionsNormalized} from './types';
 
-export const DIM_COLOR = chalk.dim;
-export const EXPECTED_COLOR = chalk.green;
-export const INVERTED_COLOR = chalk.inverse;
-export const RECEIVED_COLOR = chalk.red;
+export const INVERTED_COLOR = chalk.inverse; // export for joinAlignedDiffs test
 const PATCH_COLOR = chalk.yellow;
 
 // Given change op and array of diffs, return concatenated string:
 // * include common strings
 // * include change strings which have argument op (inverse highlight)
 // * exclude change strings which have opposite op
-export const getHighlightedString = (op: number, diffs: Array<Diff>): string =>
+export const invertChangedSubstrings = (
+  op: number,
+  diffs: Array<Diff>,
+): string =>
   diffs.reduce(
     (reduced: string, diff: Diff): string =>
       reduced +
@@ -44,14 +40,6 @@ export const getHighlightedString = (op: number, diffs: Array<Diff>): string =>
     '',
   );
 
-export const getExpectedString = (diffs: Array<Diff>): string =>
-  getHighlightedString(DIFF_DELETE, diffs);
-
-export const getReceivedString = (diffs: Array<Diff>): string =>
-  getHighlightedString(DIFF_INSERT, diffs);
-
-export const MULTILINE_REGEXP = /\n/;
-
 const NEWLINE_SYMBOL = '\u{21B5}'; // downwards arrow with corner leftwards
 const SPACE_SYMBOL = '\u{00B7}'; // middle dot
 
@@ -60,35 +48,33 @@ const SPACE_SYMBOL = '\u{00B7}'; // middle dot
 const replaceSpacesAtEnd = (line: string): string =>
   line.replace(/\s+$/, spaces => SPACE_SYMBOL.repeat(spaces.length));
 
-export const printDeleteLine = (line: string) =>
-  EXPECTED_COLOR(line.length !== 0 ? '- ' + replaceSpacesAtEnd(line) : '-');
+export const printDeleteLine = (
+  line: string,
+  {aColor, aSymbol}: DiffOptionsNormalized,
+): string =>
+  aColor(
+    line.length !== 0 ? aSymbol + ' ' + replaceSpacesAtEnd(line) : aSymbol,
+  );
 
-export const printInsertLine = (line: string) =>
-  RECEIVED_COLOR(line.length !== 0 ? '+ ' + replaceSpacesAtEnd(line) : '+');
+export const printInsertLine = (
+  line: string,
+  {bColor, bSymbol}: DiffOptionsNormalized,
+): string =>
+  bColor(
+    line.length !== 0 ? bSymbol + ' ' + replaceSpacesAtEnd(line) : bSymbol,
+  );
 
 // Prevent visually ambiguous empty line as the first or the last.
-export const printCommonLine = (line: string, isFirstOrLast: boolean = false) =>
+export const printCommonLine = (
+  line: string,
+  isFirstOrLast: boolean,
+  {commonColor, commonSymbol}: DiffOptionsNormalized,
+): string =>
   line.length !== 0
-    ? DIM_COLOR('  ' + replaceSpacesAtEnd(line))
+    ? commonColor(commonSymbol + ' ' + replaceSpacesAtEnd(line))
     : isFirstOrLast
-    ? DIM_COLOR('  ' + NEWLINE_SYMBOL)
+    ? commonColor(commonSymbol + ' ' + NEWLINE_SYMBOL)
     : '';
-
-export const computeStringDiffs = (expected: string, received: string) => {
-  const isMultiline =
-    MULTILINE_REGEXP.test(expected) || MULTILINE_REGEXP.test(received);
-
-  // getAlignedDiffs assumes that a newline was appended to the strings.
-  if (isMultiline) {
-    expected += '\n';
-    received += '\n';
-  }
-
-  const diffs = diffStrings(expected, received);
-  cleanupSemantic(diffs); // impure function
-
-  return {diffs, isMultiline};
-};
 
 export const hasCommonDiff = (diffs: Array<Diff>, isMultiline: boolean) => {
   if (isMultiline) {
@@ -102,11 +88,21 @@ export const hasCommonDiff = (diffs: Array<Diff>, isMultiline: boolean) => {
   return diffs.some(diff => diff[0] === DIFF_EQUAL);
 };
 
-export const printAnnotation = (options?: DiffOptions): string =>
-  EXPECTED_COLOR('- ' + ((options && options.aAnnotation) || 'Expected')) +
-  '\n' +
-  RECEIVED_COLOR('+ ' + ((options && options.bAnnotation) || 'Received')) +
-  '\n\n';
+export const printAnnotation = ({
+  aAnnotation,
+  aColor,
+  aSymbol,
+  bAnnotation,
+  bColor,
+  bSymbol,
+  omitAnnotationLines,
+}: DiffOptionsNormalized): string =>
+  omitAnnotationLines
+    ? ''
+    : aColor(aSymbol + ' ' + aAnnotation) +
+      '\n' +
+      bColor(bSymbol + ' ' + bAnnotation) +
+      '\n\n';
 
 // In GNU diff format, indexes are one-based instead of zero-based.
 export const createPatchMark = (
@@ -119,63 +115,88 @@ export const createPatchMark = (
     `@@ -${aStart + 1},${aEnd - aStart} +${bStart + 1},${bEnd - bStart} @@`,
   );
 
-// Return formatted diff lines without labels.
-export const printMultilineStringDiffs = (
-  diffs: Array<Diff>,
-  expand: boolean,
+// Given two string arguments, compare them character-by-character.
+// Format as comparison lines in which changed substrings have inverse colors.
+export const diffStringsUnified = (
+  a: string,
+  b: string,
+  options?: DiffOptions,
 ): string => {
-  const lines = getAlignedDiffs(diffs);
-  return expand
-    ? joinAlignedDiffsExpand(lines)
-    : joinAlignedDiffsNoExpand(lines);
+  const optionsNormalized = normalizeDiffOptions(options);
+
+  if (a.length === 0 || b.length === 0) {
+    const lines: Array<string> = [];
+
+    // All comparison lines have aColor and aSymbol.
+    if (a.length !== 0) {
+      a.split('\n').forEach(line => {
+        lines.push(printDeleteLine(line, optionsNormalized));
+      });
+    }
+
+    // All comparison lines have bColor and bSymbol.
+    if (b.length !== 0) {
+      b.split('\n').forEach(line => {
+        lines.push(printInsertLine(line, optionsNormalized));
+      });
+    }
+
+    // If both are empty strings, there are no comparison lines.
+    return printAnnotation(optionsNormalized) + lines.join('\n');
+  }
+
+  if (a === b) {
+    const lines = a.split('\n');
+    const iLast = lines.length - 1;
+
+    // All comparison lines have commonColor and commonSymbol.
+    return (
+      printAnnotation(optionsNormalized) +
+      lines
+        .map((line, i) =>
+          printCommonLine(line, i === 0 || i === iLast, optionsNormalized),
+        )
+        .join('\n')
+    );
+  }
+
+  const isMultiline = a.includes('\n') || b.includes('\n');
+
+  // getAlignedDiffs assumes that a newline was appended to the strings.
+  const diffs = diffStringsRaw(
+    isMultiline ? a + '\n' : a,
+    isMultiline ? b + '\n' : b,
+    true, // cleanupSemantic
+  );
+
+  if (hasCommonDiff(diffs, isMultiline)) {
+    const lines = getAlignedDiffs(diffs);
+    return (
+      printAnnotation(optionsNormalized) +
+      (optionsNormalized.expand
+        ? joinAlignedDiffsExpand(lines, optionsNormalized)
+        : joinAlignedDiffsNoExpand(lines, optionsNormalized))
+    );
+  }
+
+  // Fall back to line-by-line diff.
+  // Given strings, it returns a string, not null.
+  return diffLines(a, b, optionsNormalized) as string;
 };
 
-const MAX_DIFF_STRING_LENGTH = 20000;
+// Given two string arguments, compare them character-by-character.
+// Optionally clean up small common substrings, also known as chaff.
+// Return an array of diff objects.
+export const diffStringsRaw = (
+  a: string,
+  b: string,
+  cleanup: boolean,
+): Array<Diff> => {
+  const diffs = diffStrings(a, b);
 
-type StringDiffResult =
-  | {isMultiline: true; annotatedDiff: string}
-  | {isMultiline: false; a: string; b: string}
-  | null;
-
-// Print specific substring diff for strings only:
-// * if strings are not equal
-// * if neither string is empty
-// * if neither string is too long
-// * if there is a common string after semantic cleanup
-export const getStringDiff = (
-  expected: string,
-  received: string,
-  options?: DiffOptions,
-): StringDiffResult => {
-  if (
-    expected === received ||
-    expected.length === 0 ||
-    received.length === 0 ||
-    expected.length > MAX_DIFF_STRING_LENGTH ||
-    received.length > MAX_DIFF_STRING_LENGTH
-  ) {
-    return null;
+  if (cleanup) {
+    cleanupSemantic(diffs); // impure function
   }
 
-  const {diffs, isMultiline} = computeStringDiffs(expected, received);
-
-  if (!hasCommonDiff(diffs, isMultiline)) {
-    return null;
-  }
-
-  return isMultiline
-    ? {
-        annotatedDiff:
-          printAnnotation(options) +
-          printMultilineStringDiffs(
-            diffs,
-            options === undefined || options.expand !== false,
-          ),
-        isMultiline,
-      }
-    : {
-        a: getExpectedString(diffs),
-        b: getReceivedString(diffs),
-        isMultiline,
-      };
+  return diffs;
 };